# idoctor_front/src/docs/organization_service/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Organization Service API.",
        "title": "Organization Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/invitation/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "invitation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitation"
                ],
                "summary": "invitation",
                "operationId": "create-invitation",
                "parameters": [
                    {
                        "description": "invitation information",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Invitation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/ID64"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/invitation/delete_or_restore/{id}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "invitation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitation"
                ],
                "summary": "invitation",
                "operationId": "restore-invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "schema": {}
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/invitation/my": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all invitations belonging to the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitation"
                ],
                "summary": "Get my invitations",
                "operationId": "get-my-invitations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "language_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/InvitationDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/invitation/page": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetContent all invitation with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitation"
                ],
                "summary": "GetContent all invitation with pagination",
                "operationId": "get-all-invitation",
                "parameters": [
                    {
                        "type": "string",
                        "default": "1",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "10",
                        "description": "Number of items per page",
                        "name": "perpage",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "TRUE",
                            "FALSE"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "TRUE",
                            "FALSE"
                        ],
                        "name": "blocked",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "language_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "organization_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "pin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "role_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/organization_dto.InvitationPage"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/invitation/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetContent all invitation with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitation"
                ],
                "summary": "GetContent all invitation with pagination",
                "operationId": "search-invitation",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "TRUE",
                            "FALSE"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "TRUE",
                            "FALSE"
                        ],
                        "name": "blocked",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "language_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "organization_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "pin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "role_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/InvitationDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/invitation/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetContent invitation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invitation"
                ],
                "summary": "GetContent invitation by ID",
                "operationId": "get-invitation-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "language_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/InvitationDto"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ID64": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "Invitation": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "integer"
                },
                "pin": {
                    "type": "integer"
                },
                "roleId": {
                    "type": "integer"
                }
            }
        },
        "InvitationDto": {
            "type": "object",
            "properties": {
                "blockedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organization": {
                    "$ref": "#/definitions/sharedutil.JsonObject"
                },
                "organizationId": {
                    "type": "integer"
                },
                "pin": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/sharedutil.JsonObject"
                },
                "roleId": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/sharedutil.JsonObject"
                }
            }
        },
...
        "organization_dto.InvitationPage": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/InvitationDto"
                    }
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "sharedutil.JsonObject": {
            "type": "object",
            "additionalProperties": {}
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}